requirements.txt
django-taggit==2.1.0
django==4.0.3
django-localflavor==3.1
django-countries==7.3.2